fix!: split other utils into subdomain and internal exports
