+++ packages/core/package.bundle.ts
@@ -6,7 +6,10 @@ export default defineConfig(() => {
     build: {
       lib: {
         entry: {
-          index: './src/_exports/index.ts',
+          'index': './src/_exports/index.ts',
+          '_exports/_internal': './src/_exports/_internal.ts',
+          '_exports/agent': './src/_exports/agent.ts',
+          '_exports/comlink': './src/_exports/comlink.ts',
         },
       },
     },
