+++ packages/react/src/hooks/documents/useDocuments.ts
@@ -1,10 +1,5 @@
-import {
-  createGroqSearchFilter,
-  type DocumentHandle,
-  isDatasetResource,
-  type QueryOptions,
-} from '@sanity/sdk'
-import {pickProperties} from '@sanity/sdk/_internal'
+import {type DocumentHandle, isDatasetResource, type QueryOptions} from '@sanity/sdk'
+import {createGroqSearchFilter, pickProperties} from '@sanity/sdk/_internal'
 import {type SortOrderingItem} from '@sanity/types'
 import {useCallback, useMemo, useState} from 'react'
