+++ packages/react/src/hooks/users/useUser.ts
@@ -1,11 +1,5 @@
-import {
-  type GetUserOptions,
-  getUsersKey,
-  getUsersState,
-  parseUsersKey,
-  resolveUsers,
-  type SanityUser,
-} from '@sanity/sdk'
+import {type GetUserOptions, getUsersState, resolveUsers, type SanityUser} from '@sanity/sdk'
+import {getUsersKey, parseUsersKey} from '@sanity/sdk/_internal'
 import {useEffect, useMemo, useState, useSyncExternalStore, useTransition} from 'react'
 
 import {useSanityInstance} from '../context/useSanityInstance'
